// src/modules/core/clusters/md-user/pages/user/types/ManageUserPage.types.ts
type User = {
  id: string
  email: string | null
  is_verified: boolean
  status: string
  profile: Profile | null
  werk: Werk | null
  sto: Sto | null
}

type GetTableUsersMeta = {
  filter_count: number
}

export type GetTableUsersResponse = {
  data: User[]
  meta: GetTableUsersMeta
}

export type ActionButtonProps = {
  user: GetTableUsersResponse['data'][number]
}

export type Data = {
  id: string
}

export type Department = {
  id: number
  name: string
}

export type GetDropdownDepartmentsResponse = {
  data: Department[]
}

export type ResetPasswordData = Pick<GetTableUsersResponse['data'][number], 'id' | 'email'>